+++ spec/features/projects/users_can_view_a_list_of_regional_casework_services_projects_spec.rb
@@ -15,6 +15,10 @@ RSpec.feature "Viewing regional casework services projects" do
       visit completed_regional_casework_services_projects_path
 
       expect(page).to have_content(I18n.t("project.table.completed.empty"))
+
+      visit unassigned_regional_casework_services_projects_path
+
+      expect(page).to have_content(I18n.t("project.table.unassigned.empty"))
     end
   end
 
@@ -26,10 +30,12 @@ RSpec.feature "Viewing regional casework services projects" do
     end
 
     let!(:completed_regional_project) { create(:conversion_project, urn: 126041, assigned_to_regional_caseworker_team: false, completed_at: Date.yesterday) }
-    let!(:in_progress_regional_project) { create(:conversion_project, urn: 126041, assigned_to_regional_caseworker_team: false) }
+    let!(:in_progress_regional_project) { create(:conversion_project, urn: 126041, assigned_to_regional_caseworker_team: false, assigned_to: user) }
 
-    let!(:completed_project) { create(:conversion_project, urn: 121583, completed_at: Date.yesterday, assigned_to_regional_caseworker_team: true) }
-    let!(:in_progress_project) { create(:conversion_project, urn: 115652, assigned_to_regional_caseworker_team: true) }
+    let!(:completed_project) { create(:conversion_project, urn: 121583, completed_at: Date.yesterday, assigned_to_regional_caseworker_team: true, assigned_to: user) }
+    let!(:in_progress_project) { create(:conversion_project, urn: 115652, assigned_to_regional_caseworker_team: true, assigned_to: user) }
+
+    let!(:unassigned_project) { create(:conversion_project, urn: 103835, assigned_to_regional_caseworker_team: true, assigned_to: nil) }
 
     context "when signed in as a Regional caseworker" do
       let(:user) { create(:user, :caseworker) }
@@ -40,6 +46,10 @@ RSpec.feature "Viewing regional casework services projects" do
 
       scenario "they can view all in progress projects" do
         view_completed_projects
+      end
+
+      scenario "they can view unassigned projects" do
+        view_unassigned_projects
       end
     end
 
@@ -53,6 +63,10 @@ RSpec.feature "Viewing regional casework services projects" do
       scenario "they can view all in progress projects" do
         view_completed_projects
       end
+
+      scenario "they can view unassigned projects" do
+        view_unassigned_projects
+      end
     end
 
     context "when signed in as a Regional delivery officer" do
@@ -64,6 +78,10 @@ RSpec.feature "Viewing regional casework services projects" do
 
       scenario "they can view all in progress projects" do
         view_completed_projects
+      end
+
+      scenario "they can view unassigned projects" do
+        view_unassigned_projects
       end
     end
 
@@ -88,6 +106,20 @@ RSpec.feature "Viewing regional casework services projects" do
 
       within("tbody") do
         expect(page).to have_content(completed_project.urn)
+        expect(page).not_to have_content(completed_regional_project.urn)
+
+        expect(page).not_to have_content(in_progress_project.urn)
+        expect(page).not_to have_content(in_progress_regional_project.urn)
+      end
+    end
+
+    def view_unassigned_projects
+      visit unassigned_regional_casework_services_projects_path
+
+      expect(page).to have_content(I18n.t("project.regional_casework_services.unassigned.title"))
+
+      within("tbody") do
+        expect(page).not_to have_content(completed_project.urn)
         expect(page).not_to have_content(completed_regional_project.urn)
 
         expect(page).not_to have_content(in_progress_project.urn)